Reformat and cleanup all scripts

+++ Assets/Scripts/Car/UserInput/CarUserInputInterpreter.cs
@@ -12,7 +12,7 @@ namespace AutomaticParking.Car.UserInput
 
         public float InterpretAsSteeringAngle(float input) => input * carData.MaxSteeringAngle;
 
-        public bool InterpretAsBreakingState(float input) => Convert.ToBoolean(input);
+        public bool InterpretAsBreakingState(int input) => Convert.ToBoolean(input);
 
         public bool InterpretAsBreakingState(bool input) => input;
     }
